api/cloud: add support for longer access tokens
This is required to support JWT tokens. Access
tokens are sent in deregister coap message query,
thus they are part of the header and the header
is limited in size (must be <=COAP_MAX_HEADER_SIZE).

To avoid this problem we don't use access token
in the deregister query. Instead we examine the
cloud connection state.
  * If the device is logged in then the token is not
  needed and we sent the deregister message without it.
  * If the device is not logged in then we try logging
  in and then we sent the deregister message without
  token.